secure/requirements.txt
action-msgs==1.2.1
action-tutorials-interfaces==0.20.3
action-tutorials-py==0.20.3
actionlib-msgs==4.2.3
ament-cmake-test==1.3.3
ament-copyright==0.12.5
ament-cppcheck==0.12.5
ament-cpplint==0.12.5
ament-flake8==0.12.5
ament-index-python==1.4.0
ament-lint==0.12.5
ament-lint-cmake==0.12.5
ament-package==0.14.0
ament-pep257==0.12.5
ament-uncrustify==0.12.5
ament-xmllint==0.12.5
angles==1.15.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.2
bleach==6.0.0
builtin-interfaces==1.2.1
camera-calibration==3.0.0
cffi==1.15.1
comm==0.1.2
composition-interfaces==1.2.1
cv-bridge==3.2.1
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
demo-nodes-py==0.20.3
diagnostic-msgs==4.2.3
Django==4.1.6
django-recaptcha==3.0.0
domain-coordinator==0.10.0
dynamixel-sdk-custom-interfaces==3.7.60
example-interfaces==0.9.3
examples-rclpy-executors==0.15.1
examples-rclpy-minimal-action-client==0.15.1
examples-rclpy-minimal-action-server==0.15.1
examples-rclpy-minimal-client==0.15.1
examples-rclpy-minimal-publisher==0.15.1
examples-rclpy-minimal-service==0.15.1
examples-rclpy-minimal-subscriber==0.15.1
executing==1.2.0
fastjsonschema==2.16.2
fqdn==1.5.1
geometry-msgs==4.2.3
idna==3.4
image-geometry==3.2.1
interactive-markers==2.3.2
ipykernel==6.21.1
ipython==8.9.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter_client==8.0.2
jupyter_core==5.2.0
jupyter_server==2.2.1
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
laser-geometry==2.4.0
launch==1.0.4
launch-ros==0.19.4
launch-testing==1.0.4
launch-testing-ros==0.19.4
launch-xml==1.0.4
launch-yaml==1.0.4
lifecycle-msgs==1.2.1
logging-demo==0.20.3
map-msgs==2.1.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
message-filters==4.3.2
mistune==2.0.4
mobilechelonian==0.5
nav-msgs==4.2.3
nbclassic==0.5.1
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
osrf-pycommon==2.0.2
packaging==23.0
pandocfilters==1.5.0
parso==0.8.3
pcl-msgs==1.0.0
pendulum-msgs==0.20.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.6.2
prometheus-client==0.16.0
prompt-toolkit==3.0.36
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.14.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.4
python-qt-binding==1.1.1
PyYAML==6.0
pyzmq==25.0.0
qt-dotgraph==2.2.2
qt-gui==2.2.2
qt-gui-cpp==2.2.2
qt-gui-py-common==2.2.2
quality-of-service-demo-py==0.20.3
rcl-interfaces==1.2.1
rclpy==3.3.7
rcutils==5.1.2
resource-retriever==3.1.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rmw-dds-common==1.6.0
ros-gz-interfaces==0.244.9
ros-ign-interfaces==0.244.9
ros2action==0.18.5
ros2bag==0.15.4
ros2cli==0.18.5
ros2component==0.18.5
ros2doctor==0.18.5
ros2interface==0.18.5
ros2launch==0.19.4
ros2lifecycle==0.18.5
ros2multicast==0.18.5
ros2node==0.18.5
ros2param==0.18.5
ros2pkg==0.18.5
ros2run==0.18.5
ros2service==0.18.5
ros2topic==0.18.5
rosbag2-interfaces==0.15.4
rosbag2-py==0.15.4
rosgraph-msgs==1.2.1
rosidl-adapter==3.1.4
rosidl-cli==3.1.4
rosidl-cmake==3.1.4
rosidl-generator-c==3.1.4
rosidl-generator-cpp==3.1.4
rosidl-generator-py==0.14.4
rosidl-parser==3.1.4
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.0
rosidl-typesupport-cpp==2.0.0
rosidl-typesupport-fastrtps-c==2.2.0
rosidl-typesupport-fastrtps-cpp==2.2.0
rosidl-typesupport-introspection-c==3.1.4
rosidl-typesupport-introspection-cpp==3.1.4
rpyutils==0.2.1
rqt-action==2.0.1
rqt-bag==1.1.4
rqt-bag-plugins==1.1.4
rqt-console==2.0.2
rqt-graph==1.3.0
rqt-gui==1.1.4
rqt-gui-py==1.1.4
rqt-msg==1.2.0
rqt-plot==1.1.2
rqt-publisher==1.5.0
rqt-py-common==1.1.4
rqt-py-console==1.0.2
rqt-reconfigure==1.1.1
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-topic==1.5.0
Send2Trash==1.8.0
sensor-msgs==4.2.3
sensor-msgs-py==4.2.3
shape-msgs==4.2.3
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
sros2==0.10.4
stack-data==0.6.2
statistics-msgs==1.2.1
std-msgs==4.2.3
std-srvs==4.2.3
stereo-msgs==4.2.3
teleop-twist-keyboard==2.3.2
terminado==0.17.1
tf2-geometry-msgs==0.25.2
tf2-kdl==0.25.2
tf2-msgs==0.25.2
tf2-py==0.25.2
tf2-ros-py==0.25.2
tf2-tools==0.25.2
theora-image-transport==2.5.0
tinycss2==1.2.1
topic-monitor==0.20.3
tornado==6.2
traitlets==5.9.0
trajectory-msgs==4.2.3
# Editable install with no version control (turtlebot3-example==2.1.5)
-e /home/ulbert/turtlebot3_ws/build/turtlebot3_example
turtlebot3-msgs==2.2.3
# Editable install with no version control (turtlebot3-teleop==2.1.5)
-e /home/ulbert/turtlebot3_ws/build/turtlebot3_teleop
turtlesim==1.4.2
unique-identifier-msgs==2.2.1
uri-template==1.2.0
visualization-msgs==4.2.3
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
widgetsnbextension==4.0.5
xacro==2.0.8
